pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py>=4.8.0",
    "langchain>=1.0.5",
    "langchain-classic>=1.0.0",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "langchain-deepseek>=1.0.1",
    "langchain-community>=0.4.1",
    "python-pptx>=1.0.2",
    "langchain-anthropic>=1.1.0",
    "langchain-deepseek>=1.0.1",
    "python-dotenv>=1.2.1",
    "firecrawl>=4.9.0",
]

